# backend/swagger.yaml
openapi: 3.0.0
info:
  title: M15ASushi - Invoice Creation and Validation API
  description: |
    A comprehensive API for creating UBL 2.4-compliant invoices and validating against PEPPOL standards. 
    
    ## v1 capabilities:
    - Generate UBL 2.4-compliant invoices from JSON objects
    - Validate UBL-XML invoice against Australian and New Zealand PEPPOL standards
    - Secure storage of invoices requiring specific user authentication to access
    - Retrieve invoices in XML format
    
    # Guide to testing:
    
    We strongly recommend using [Postman](https://www.postman.com/) for testing our API as it will allow you to authenticate requests easiest.
    
    1. Copy and paste the following script in **Scripts->Post-response**.
          ```javascript var jsonData = pm.response.json();
          if (jsonData.status === "success" && jsonData.data && jsonData.data.token) {
              pm.environment.set("token", jsonData.data.token);
          }
    2. Register or login using the **/v1/users/register** or **/v1/users/login** endpoints respectively.
    3. Copy the `token`, then go to **Authorization**, select **Bearer Token** for the Auth Type and paste your `token`.
    4. All done! Your token will now be automatically used when you call our routes :)
  version: 1.0.0
servers:
  - url: http://sushi-invoice-application.ap-southeast-2.elasticbeanstalk.com
paths:
...
  /v1/invoices/create:
    post:
      summary: Create a new invoice
      description: |
        This endpoint allows the creation of a new invoice with various fields that are inputted as a JSON object and will be converted to UBL XML format. Each field will be checked for its type and format (see the schema for more details). Required fields are: 
          - invoiceId
          - issueDate
          - buyer
          - supplier
          - total
          - items
      operationId: createInvoice
      tags:
        - Invoices
      parameters:
        - name: token
          in: cookie
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: Feel free to copy this request body for testing.
        content:
          application/json:
            schema:
              type: object
              properties:
                invoiceId:
                  type: string
                  example: TEST-INVOICE-ID
                supplier:
                  type: string
                  example: supplier name
                buyer:
                  type: string
                  example: buyer name
                total:
                  type: number
                  format: double
                  example: 9.99
                  description: Must be equal the sum of the item cost(s)
                currency:
                  type: string
                  description: Must be a valid 3 letter currency code (ISO 4217)
                  example: "AUD"
                issueDate:
                  type: string
                  format: date
                  description: Must be a past date and be of the format YYYY-MM-DD
                dueDate:
                  type: string
                  format: date
                  description: Must be on or after the issue date and be of the format YYYY-MM-DD
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      name: 
                        type: string
                      count:
                        type: number
                        example: 1
                        description: Must be greater than 0
                      cost:
                        type: number
                        format: double
                        example: 9.99
                buyerAddress:                                    
                  type: object
                  properties:
                    street:
                      type: string
                      description: Street address of the buyer
                    country:
                      type: string
                      description: Country code of the buyer (ISO 3166-1 alpha-2)
                      example: "AU"
                buyerEmail:
                  type: string
                  format: email
                  description: Email address of the buyer
                buyerPhone:
                  type: string
                  description: Phone number of the buyer
                supplierAddress:
                  type: object
                  properties:
                    street:
                      type: string
                      description: Street address of the supplier
                    country:
                      type: string
                      description: Country code of the supplier (ISO 3166-1 alpha-2)
                      example: "AU"
                supplierEmail:
                  type: string
                  format: email
                  description: Email address of the supplier
                supplierPhone:
                  type: string
                  description: Phone number of the supplier
                paymentAccountId:
                  type: string
                  description: Bank account number
                paymentAccountName:
                  type: string
                  description: Account name
                financialInstitutionBranchId:
                  type: string
                  description: Branch identifier such as BSB
      responses:
        "200":
          description: Invoice created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceid:
                    type: string
        "400":
          description: If there are any missing requried fields or invalid details provided in the request body. 
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
        "401":
          description: Invalid token
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
...
  /v1/invoices/{invoiceid}/validate:
    post:
      summary: Validate invoice standard
      description: |
        This endpoint validates an invoice in the database against the UBL 2.4 standard as well as Australian Peppol standards.
        
        If the invoice is invalid, it will return a list of validation errors, with information on what is missing for a valid infvoice. 
        
        A list of checks performed is as follows (users only has to worry about points in **bold**):
        - Has the correct namesapce tag
        - Has all requried UBL tags: 'cbc:ID', 'cbc:IssueDate', 'cbc:InvoiceTypeCode', 'cac:AccountingSupplierParty', 'cac:AccountingCustomerParty', 'cac:LegalMonetaryTotal', 'cac:InvoiceLine', 'cbc:CustomizationID', 'cbc:ProfileID', 'cac:OrderReference', 'cac:InvoiceDocumentReference'
        - Valid invoice type code
        - **Buyer and supplier information is included (name, address, and country code)**
        - **Valid currency and country codes**
        - Items have necessary information (ID, **name, price, count**)
        
         
        
      operationId: validateInvoice
      tags:
        - Invoices
      parameters:
        - name: invoiceid
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: cookie
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice validated successfully
        "400":
          description: |
            Invalid details: failed one or more of the checks
          content: 
            application/json:
              schema:
                type: object
                properties:
                  error: 
                    type: array
                    items:
                      type: string
                      example: 'Missing ____ (Peppol rule ___)'
        "401":
          description: Invalid token
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
        "403":
          description: Unauthorised access
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
  /v1/invoices/create-and-validate:
    post:
      summary: Create and validate an invoice
      description: |
        This endpoint has the same function as calling both the /create and /validate endopoints. It creates a new invoice then validates it against the UBL 2.4 standard as well as Australian Peppol standards.
        
        Note: The invoice will still be created even if the validation fails. 
      operationId: createAndValidateInvoice
      tags:
        - Invoices
      parameters:
        - name: token
          in: cookie
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: Feel free to copy this request body for testing.
        content:
          application/json:
            schema:
              type: object
              properties:
                invoiceId:
                  type: string
                  example: TEST-INVOICE-ID
                supplier:
                  type: string
                  example: supplier name
                buyer:
                  type: string
                  example: buyer name
                total:
                  type: number
                  format: double
                  example: 9.99
                  description: Must be equal the sum of the item cost(s)
                currency:
                  type: string
                  description: Must be a valid 3 letter currency code (ISO 4217)
                  example: "AUD"
                issueDate:
                  type: string
                  format: date
                  description: Must be a past date and be of the format YYYY-MM-DD
                dueDate:
                  type: string
                  format: date
                  description: Must be on or after the issue date and be of the format YYYY-MM-DD
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      name: 
                        type: string
                      count:
                        type: number
                        example: 1
                        description: Must be greater than 0
                      cost:
                        type: number
                        format: double
                        example: 9.99
                buyerAddress:                                    
                  type: object
                  properties:
                    street:
                      type: string
                      description: Street address of the buyer
                    country:
                      type: string
                      description: Country code of the buyer (ISO 3166-1 alpha-2)
                      example: "AU"
                buyerEmail:
                  type: string
                  format: email
                  description: Email address of the buyer
                buyerPhone:
                  type: string
                  description: Phone number of the buyer
                supplierAddress:
                  type: object
                  properties:
                    street:
                      type: string
                      description: Street address of the supplier
                    country:
                      type: string
                      description: Country code of the supplier (ISO 3166-1 alpha-2)
                      example: "AU"
                supplierEmail:
                  type: string
                  format: email
                  description: Email address of the supplier
                supplierPhone:
                  type: string
                  description: Phone number of the supplier
                paymentAccountId:
                  type: string
                  description: Bank account number
                paymentAccountName:
                  type: string
                  description: Account name
                financialInstitutionBranchId:
                  type: string
                  description: Branch identifier such as BSB
      responses:
        "200":
          description: Invoice updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                    example: INV-001
        "400":
          description: Creation or validation failure due to invalid input
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
        "401":
          description: Invalid token
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
...